refactor: move HttpHelper to a folder
