+++ src/main/decorators/log.spec.ts
@@ -1,13 +1,13 @@
 import { LogControllerDecorator } from './log'
 import { Controller, HttpRequest, HttpResponse } from '../../presentation/protocols'
-import { serverError, ok } from '../../presentation/helpers/http-helper'
 import { LogErrorRepository } from '../../data/protocols/log-error-repository'
 import { AccountModel } from '../../domain/models/account'
+import { ok, serverError } from '../../presentation/helpers/http/http-helper'
 
 const makeController = (): Controller => {
   class ControllerStub implements Controller {
-    async handle (httpRequest: HttpRequest): Promise<HttpResponse> {
-      return new Promise(resolve => resolve(ok(makeFakeAccount())))
+    async handle(httpRequest: HttpRequest): Promise<HttpResponse> {
+      return new Promise((resolve) => resolve(ok(makeFakeAccount())))
     }
   }
   return new ControllerStub()
@@ -15,8 +15,8 @@ const makeController = (): Controller => {
 
 const makeLogErrorRepository = (): LogErrorRepository => {
   class LogErrorRepositoryStub implements LogErrorRepository {
-    async logError (stack: string): Promise<void> {
-      return new Promise(resolve => resolve())
+    async logError(stack: string): Promise<void> {
+      return new Promise((resolve) => resolve())
     }
   }
   return new LogErrorRepositoryStub()
@@ -27,15 +27,15 @@ const makeFakeRequest = (): HttpRequest => ({
     name: 'any_name',
     email: 'any_email@mail.com',
     password: 'any_password',
-    passwordConfirmation: 'any_password'
-  }
+    passwordConfirmation: 'any_password',
+  },
 })
 
 const makeFakeAccount = (): AccountModel => ({
   id: 'valid_id',
   name: 'valid_name',
   email: 'valid_email@mail.com',
-  password: 'valid_password'
+  password: 'valid_password',
 })
 
 const makeFakeServerError = (): HttpResponse => {
@@ -57,7 +57,7 @@ const makeSut = (): SutTypes => {
   return {
     sut,
     controllerStub,
-    logErrorRepositoryStub
+    logErrorRepositoryStub,
   }
 }
 
@@ -78,7 +78,7 @@ describe('LogController Decorator', () => {
   test('Should call LogErrorRepository with correct error if controller returns a server error', async () => {
     const { sut, controllerStub, logErrorRepositoryStub } = makeSut()
     const logSpy = jest.spyOn(logErrorRepositoryStub, 'logError')
-    jest.spyOn(controllerStub, 'handle').mockReturnValueOnce(new Promise(resolve => resolve(makeFakeServerError())))
+    jest.spyOn(controllerStub, 'handle').mockReturnValueOnce(new Promise((resolve) => resolve(makeFakeServerError())))
     await sut.handle(makeFakeRequest())
     expect(logSpy).toHaveBeenCalledWith('any_stack')
   })
