+++ src/presentation/controllers/login/login.spec.ts
@@ -1,5 +1,5 @@
 import { HttpRequest, Authentication, Validation } from './login-protocols'
-import { badRequest, ok, serverError, unauthorized } from '../../helpers/http-helper'
+import { badRequest, ok, serverError, unauthorized } from '../../helpers/http/http-helper'
 import { MissingParamError } from '../../errors'
 
 import { LoginController } from './login'
